requirements.txt
langchain==0.1.16
langchain-google-genai==0.0.11
python-dotenv==1.0.1
requests==2.31.0
